add preloader to all pages
