+++ server/__tests__/mockData.js
@@ -88,7 +88,7 @@ const mockData = {
   },
   document: {
     title: faker.lorem.words(2),
-    synopsis: faker.lorem.sentences(),
+    synopsis: faker.lorem.sentences(100),
     body: faker.lorem.sentences(),
     owner: 4,
     accessRight: 'private',
