+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const webpack = require('webpack');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
-
+const CompressionPlugin = require('compression-webpack-plugin');
 
 const debug = process.env.NODE_ENV !== 'production';
 const basePlugins = [
@@ -26,10 +26,17 @@ const basePlugins = [
 const debugPlugins = [new ExtractTextPlugin('style.css')];
 const productionPlugins = [
   new webpack.optimize.OccurrenceOrderPlugin(),
-  new webpack.optimize.UglifyJsPlugin({ mangle: false, sourcemap: false }),
+  new webpack.optimize.UglifyJsPlugin(),
   new ExtractTextPlugin({
     filename: path.join(__dirname, 'public/style.css'),
     allChunks: true
+  }),
+  new CompressionPlugin({
+    asset: '[path].gz[query]',
+    algorithm: 'gzip',
+    test: /\.js$|\.css$|\.html$/,
+    threshold: 10240,
+    minRatio: 0.8
   })
 ];
